University Gym Usage Analysis
The main reason for analyzing this data is to help the university with the staffing, facilities and open hour modifications. Apart from that it also enables the students to find the best possible time for them to go to the gym depending on their preference of how crowded the gym is.